@@ -19,15 +19,15 @@
     "@formatjs/intl-locale": "^3.3.2",
     "@formatjs/intl-localematcher": "^0.4.0",
     "@formatjs/intl-pluralrules": "^5.2.4",
-    "@shopify/app-bridge": "^3.7.7",
-    "@shopify/app-bridge-react": "^3.7.7",
+    "@shopify/app-bridge": "^3.7.10",
+    "@shopify/app-bridge-react": "^3.7.10",
     "@shopify/i18next-shopify": "^0.2.9",
-    "@shopify/polaris": "^10.49.1",
+    "@shopify/polaris": "^12.9.1",
     "@vitejs/plugin-react": "1.2.0",
     "i18next": "^23.1.0",
     "i18next-resources-to-backend": "^1.1.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^13.0.0",
     "react-query": "^3.34.19",
     "react-router-dom": "^6.3.0",
