@@ -1,3 +1,10 @@
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+
+const __dirname = path.dirname(__filename);
+
 export const DEFAULT_FETCH_DELAY = 1000; // in ms
 export const MAX_RETRY_AMOUNT = 100; // MAX_RETRY_AMOUNT * DEFAULT_FETCH_DELAY = max time to wait for response
 
@@ -14,5 +21,7 @@ export const WEBHOOKS = [
 
 export const STATIC_PATH =
   process.env.NODE_ENV === 'production'
-    ? `${process.cwd()}/frontend/dist`
-    : `${process.cwd()}/frontend/`;
+    ? path.resolve(__dirname, '..', '..', '..', 'frontend/dist') //; `${process.cwd()}/frontend/dist`
+    : path.resolve(__dirname, '..', '..', '..', 'frontend'); //`${process.cwd()}/frontend/`;
+
+console.log('STATIC_PATH', process.env.NODE_ENV, STATIC_PATH);
